+++ src/main/java/com/example/sweet/database/schema/Loai/TanSuatNhanLai.java
@@ -8,18 +8,18 @@ import jakarta.persistence.Id;
 @Entity
 public class TanSuatNhanLai {
     @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private int tanSoNhanLaiID;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long tanSoNhanLaiID;
     private String tenTanSoNhanLai;
     private int maTanSoNhanLai;
     private String moTa;
     private boolean coHoatDong;
 
-    public int getTanSoNhanLaiID() {
+    public Long getTanSoNhanLaiID() {
         return tanSoNhanLaiID;
     }
 
-    public void setTanSoNhanLaiID(int tanSoNhanLaiID) {
+    public void setTanSoNhanLaiID(Long tanSoNhanLaiID) {
         this.tanSoNhanLaiID = tanSoNhanLaiID;
     }
